Create Trendline & Support indicator
This indicator can be very useful when doing technical analysis regardless of experience. The lines on the chart simply display approximate support and resistance zones based off repeated price action. In terms of the trend lines, a fibonacci extension channel is plotted automatically onto the chart to provide some context predictions of the trend direction.