Create RSI, Stoch Rsi, EMA, SMA, & ROC
This indicator is simply an enhanced version of the RSI followed up by a few extra indicators that pair strongly with the RSI . This indicator allows the user to interact with various inputs based off the indicators provided. All indicators include moving average, relative strength index , stochastic relative strength index , simple moving average , exponential moving average , and rate of change . This program is unique as it is very versatile allowing the user to use as little or as many indicators as needed interchangeably.